flask_app/requirements.txt
certifi==2019.6.16
Click==7.0
cloudpickle==1.2.1
cycler==0.10.0
cytoolz==0.10.0
dask==2.2.0
decorator==4.4.0
Flask==1.1.1
Flask-Uploads==0.2.1
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
networkx==2.3
numpy==1.17.0
olefile==0.46
opencv-python==4.1.0.25
Pillow==6.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
toolz==0.10.0
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
Werkzeug==0.15.4
wincertstore==0.2
